refactor: move HD/PRB algorithms to core_algorithm
